nhan

+++ service02/src/main/java/com/example/service02/javax1/model/user/User.java
@@ -21,7 +21,7 @@ public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    protected long id;
+    protected int id;
 
     @NotEmpty(message = "Name is required!")
     @NotBlank
